Reorganised codebase with modularisation
Generic functions are stored in a file, each. All files are stored in the folder utility. Aggregation is done via a barrel.js file. Two procedures are run concurrently: one for generating a hierarchical representation of a file tree, the other to inject the Excel table in a JSON. Then, both are merged to feature the parent property.